fix(functions): Truncate fingerprint to uint32
The fingerprint is being truncated to uint32 at read time already, so we should
trunate at write time as well. (see getsentry/sentry#52475)

+++ internal/nodetree/nodetree.go
@@ -171,7 +171,7 @@ func (n *Node) CollectFunctions(profilePlatform platform.Platform, results map[u
 			function, exists := results[fingerprint]
 			if !exists {
 				results[fingerprint] = CallTreeFunction{
-					Fingerprint:   fingerprint,
+					Fingerprint:   uint64(uint32(fingerprint)),
 					Function:      n.Frame.Function,
 					Package:       framePackage,
 					InApp:         n.IsApplication,
